% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sydney_water_temp}
\alias{sydney_water_temp}
\title{NSW Beachwatch Water Quality Data}
\format{
A data frame with 8947 rows and 7 columns:
\describe{
\item{temp}{Water Temperature in C}
\item{beach}{Water measurement site}
\item{date}{Measurement date}
\item{time}{Measurement time}
\item{month}{Numeric label for month of measurement}
\item{hour}{Numeric label for hour of measurement}
\item{month_lab}{Factor label for month of measurement}
...
}
}
\source{
\url{https://beachwatch.nsw.gov.au/waterMonitoring/waterQualityData}

licensed under the Creative Commons Attribution 4.0 International (CC BY 4.0)
}
\usage{
sydney_water_temp
}
\description{
Water temperature and quality data collected under the NSW Government Beachwatch program.
}
\keyword{datasets}
